feat: add sdk uniqueness headers (#101)
Updates the dependency on the underlying Unleash client to 3.7.2. This
version of the client adds new x-unleash headers which lets Unleash
count unique connections and see SDK version information.

Also updates the node client to 6.4.1 which includes the new headers for the server side (even though next uses it's own fetching).